// src/app/interfaces/pictogram.interface.ts
// Generated by https://quicktype.io

export interface Pictogram {
    pictoImageUrl: string;
    schematic:   boolean;
    sex:         boolean;
    violence:    boolean;
    aac:         boolean;
    aacColor:    boolean;
    skin:        boolean;
    hair:        boolean;
    downloads:   number;
    categories:  string[];
    synsets:     string[];
    tags:        string[];
    _id:         number;
    created:     string;
    lastUpdated: string;
    keywords:    Keyword[];
    desc:        string;
}

export interface Keyword {
    keyword:     string;
    type:        number;
    meaning:     string;
    plural:      string;
    hasLocution: boolean;
}